+++ test/uuidtrip.ml
@@ -51,10 +51,10 @@ let version =
 let ns =
   let ns_arg =
     let parse s = match Uuidm.of_string s with
-    | None -> `Error (strf "%S: could not parse namespace UUID" s)
-    | Some ns -> `Ok ns
+    | None -> Error (strf "%S: could not parse namespace UUID" s)
+    | Some ns -> Ok ns
     in
-    parse, Uuidm.pp
+    Arg.conv' ~docv:"UUID" (parse, Uuidm.pp)
   in
   let doc = "Namespace UUID for name based UUIDs (version 4 or 5).
              Defaults to the DNS namespace UUID."